requirements.txt
# Core scraper dependencies
pandas>=1.3.3
requests>=2.26.0
selenium>=4.7.2
selenium-wire>=4.5.4
webdriver-manager>=3.5.1
xlsxwriter>=3.1.9
python-dotenv>=0.19.0

# Web app dependencies
streamlit>=1.28.0
plotly>=5.0.0 